pyproject.toml
[project]
name = "wchw-green-agent"
version = "1.0.0"
description = "Green Agent for WCHW (Wireless Communication Homework) Benchmark - UC Berkeley AgentBeats Competition"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    {name = "Jingwen Tong", email = "jingwen.tong@example.com"}
]
keywords = [
    "a2a",
    "agent",
    "wireless-communication",
    "benchmark",
    "evaluation",
    "agentbeats"
]
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "pydantic>=2.12.5",
    "uvicorn>=0.38.0",
    "aiofiles>=24.1.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.28.1",
]
